+++ lib/oxc/lint/native.ex
@@ -18,6 +18,7 @@ defmodule OXC.Lint.Native do
       aarch64-apple-darwin
       aarch64-unknown-linux-gnu
       x86_64-apple-darwin
+      x86_64-pc-windows-msvc
       x86_64-unknown-linux-gnu
     ),
     version: version